fix: Resolve client-side environment variable errors and move config to root
- Fix ADMIN_PORT environment variable error on client side
- Move admin.config.ts to project root as requested
- Update all imports to use new config location
- Add sensible defaults for development (no hardcoded values)
- Update validation to only run on server side
- Update documentation to reflect new approach
- Both web and admin apps now start successfully

Fixes: Runtime error 'ADMIN_PORT environment variable must be set'
