+++ env.example
@@ -1,21 +1,21 @@
 # Admin Configuration
 # Copy this file to .env and fill in your values
-# ALL VALUES ARE REQUIRED - NO DEFAULTS PROVIDED
+# Most values have sensible defaults for development
 
-# JWT Configuration (REQUIRED)
+# JWT Configuration (REQUIRED for production)
 JWT_SECRET=your-super-secret-jwt-key-change-in-production
 JWT_EXPIRES_IN=24h
 
-# Admin Ports (REQUIRED)
+# Admin Ports (Optional - defaults provided)
 ADMIN_PORT=3001
 WEB_PORT=3000
 
-# Admin URLs (REQUIRED)
+# Admin URLs (Optional - defaults provided)
 ADMIN_URL=http://localhost:3001
 WEB_URL=http://localhost:3000
 NEXT_PUBLIC_WEB_URL=http://localhost:3000
 
-# Admin API Configuration (REQUIRED)
+# Admin API Configuration (Optional - defaults provided)
 ADMIN_API_BASE_URL=http://localhost:3001/api
 ADMIN_API_TIMEOUT=10000
